add exception throw in delete method of service and test it successfully

+++ backend/src/main/java/org/example/backend/VocabService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @RequiredArgsConstructor
 @Service
@@ -19,6 +20,8 @@ public class VocabService {
     }
 
     public String deleteVocab(String id){
+       if ( !vocabRepo.existsById(id)) {throw new NoSuchElementException();
+       }
         vocabRepo.deleteById(id);
         return "Vocab successfully deleted.";
     }
